+++ app/models/concerns/folio/friendly_id.rb
@@ -19,10 +19,7 @@ module Folio::FriendlyId
     else
       friendly_id :slug_candidates, use: %i[slugged history]
 
-      validates :slug,
-                presence: true,
-                uniqueness: true,
-                format: { with: /[0-9a-z-]+/ }
+      validates :slug, presence: true, uniqueness: true, format: { with: /[0-9a-z-]+/ }
     end
 
     before_validation :strip_and_downcase_slug